added custom icon

+++ ImportCSVToSharePoint.ps1
@@ -1,27 +1,27 @@
 #install ad module from here: https://www.microsoft.com/en-us/download/details.aspx?id=45520
 Connect-PnPOnline -Url "https://garaioag874.sharepoint.com/sites/SPFxShowcase"
 cd D:\AVexcluded\WeeklyResourcePlanning
-Remove-PnPList "RessourcenPlanDaten" -Force
-New-PnPList -Title "RessourcenPlanDaten" -Template GenericList
-New-PnPList -Title "ProjectInformation" -Template GenericList
 
-$RessourcenPlanDatenList = Get-PnPList -Identity "RessourcenPlanDaten"
-$projectInformationList = Get-PnPList -Identity "ProjectInformation"
+if ($false) {
+    Remove-PnPList "RessourcenPlanDaten" -Force
+    New-PnPList -Title "RessourcenPlanDaten" -Template GenericList
+    New-PnPList -Title "ProjectInformation" -Template GenericList
 
-Add-PnPField -List $projectInformationList -DisplayName "ProjectSpaceRelativeUrl" -InternalName "ProjectSpaceRelativeUrl" -Type URL -AddToDefaultView
-Add-PnPField -List $projectInformationList -DisplayName "JiraAbsoluteUrl" -InternalName "JiraAbsoluteUrl" -Type URL -AddToDefaultView
+    $RessourcenPlanDatenList = Get-PnPList -Identity "RessourcenPlanDaten"
+    $projectInformationList = Get-PnPList -Identity "ProjectInformation"
 
-$projectInformationListItems = Get-PnPListItem -List "ProjectInformation" -Fields "Title", "ProjectSpaceRelativeUrl", "JiraAbsoluteUrl"
+    Add-PnPField -List $projectInformationList -DisplayName "ProjectSpaceRelativeUrl" -InternalName "ProjectSpaceRelativeUrl" -Type URL -AddToDefaultView
+    Add-PnPField -List $projectInformationList -DisplayName "JiraAbsoluteUrl" -InternalName "JiraAbsoluteUrl" -Type URL -AddToDefaultView
 
-#ProjektCode field not needed. Rename title field to ProjektCode
-#Add-PnPField -List $RessourcenPlanDatenList -DisplayName "ProjektCode" -InternalName "ProjektCode" -Type Text -AddToDefaultView
-Add-PnPField -List $RessourcenPlanDatenList -DisplayName "PlanMinuten" -InternalName "PlanMinuten" -Type Number -AddToDefaultView
-Add-PnPField -List $RessourcenPlanDatenList -DisplayName "IstMinuten" -InternalName "IstMinuten" -Type Number -AddToDefaultView
-Add-PnPField -List $RessourcenPlanDatenList -DisplayName "ProjectSpaceRelativeUrl" -InternalName "ProjectSpaceRelativeUrl" -Type URL -AddToDefaultView
-Add-PnPField -List $RessourcenPlanDatenList -DisplayName "JiraAbsoluteUrl" -InternalName "JiraAbsoluteUrl" -Type URL -AddToDefaultView
-Add-PnPField -List $RessourcenPlanDatenList -DisplayName "User" -InternalName "User" -Type User -AddToDefaultView
+    #ProjektCode field not needed. Rename title field to ProjektCode
+    #Add-PnPField -List $RessourcenPlanDatenList -DisplayName "ProjektCode" -InternalName "ProjektCode" -Type Text -AddToDefaultView
+    Add-PnPField -List $RessourcenPlanDatenList -DisplayName "PlanMinuten" -InternalName "PlanMinuten" -Type Number -AddToDefaultView
+    Add-PnPField -List $RessourcenPlanDatenList -DisplayName "IstMinuten" -InternalName "IstMinuten" -Type Number -AddToDefaultView
+    Add-PnPField -List $RessourcenPlanDatenList -DisplayName "ProjectSpaceRelativeUrl" -InternalName "ProjectSpaceRelativeUrl" -Type URL -AddToDefaultView
+    Add-PnPField -List $RessourcenPlanDatenList -DisplayName "JiraAbsoluteUrl" -InternalName "JiraAbsoluteUrl" -Type URL -AddToDefaultView
+    Add-PnPField -List $RessourcenPlanDatenList -DisplayName "User" -InternalName "User" -Type User -AddToDefaultView
 
-Add-PnPFieldFromXml '<Field Type="DateTime"
+    Add-PnPFieldFromXml '<Field Type="DateTime"
 				DisplayName="WochenDatum"
 				Required="FALSE"
 				EnforceUniqueValues="FALSE"
@@ -33,10 +33,38 @@ Add-PnPFieldFromXml '<Field Type="DateTime"
 				StaticName="WochenDatum"
 				Name="WochenDatum">
     </Field>' -List $RessourcenPlanDatenList
-#add DateTime field to default view (not possible per PS yet)
+    #add DateTime field to default view (not possible per PS yet)
+}
+else {
+    $RessourcenPlanDatenList = Get-PnPList -Identity "RessourcenPlanDaten"
+    $projectInformationList = Get-PnPList -Identity "ProjectInformation"
+}
+
+$projectInformationListItems = Get-PnPListItem -List "ProjectInformation" -Fields "Title", "ProjectSpaceRelativeUrl", "JiraAbsoluteUrl"
 
 Invoke-WebRequest -Uri "https://reporting.garaio.com/ReportServer/Pages/ReportViewer.aspx?/VertecReports/RessourcenPlanDaten&rs:Format=CSV" -OutFile ./RessourcenPlanDaten.csv -UseDefaultCredentials
 $import = Import-Csv -Delimiter "," -Path "RessourcenPlanDaten.csv"
+
+#add new projects to projectInformation list
+$existingProjects = @()
+$projectInformationListItems | % {
+    $item = New-Object PSObject
+    $item | Add-Member -type NoteProperty -Name 'ProjektCode' -Value $_.FieldValues.Title
+    $existingProjects += $item
+}
+$newProjects = $import | select ProjektCode -Unique
+if ($existingProjects -ne $null) {
+    $projectsToAdd = Compare-Object -Property ProjektCode -ReferenceObject $existingProjects -DifferenceObject $newProjects | ? {$_.SideIndicator -eq "=>"} | select ProjektCode
+}else {
+    $projectsToAdd = $newProjects
+}
+
+$projectsToAdd | % {
+    $values = @{
+        "Title" = $_.ProjektCode;
+    }
+    Add-PnPListItem -List $projectInformationList -Values $values -ContentType Element
+}
 
 $import | % {
     $projektCode = $_.ProjektCode
